backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pytesseract
pillow
langchain
langchain-community
langchain-groq
langchain-google-genai
faiss-cpu
python-dotenv
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
reportlab
langchain-text-splitters
